Add plumbing for Pin from V1
