+++ gazelle/internal/spreso/pin_store_v1_test.go
@@ -1,0 +1,7 @@
+package spreso_test
+
+import "testing"
+
+func TestNewPinsFromV1PinStore(t *testing.T) {
+	t.Error("IMPLEMENT ME!")
+}
