+++ gazelle/internal/spreso/pin_store_v2_test.go
@@ -1,0 +1,7 @@
+package spreso_test
+
+import "testing"
+
+func TestNewPinsFromV2PinStore(t *testing.T) {
+	t.Error("IMPLEMENT ME!")
+}
